Escape Room code completed
This is a board game called "Escape Room", in which the player must escape from a locked room. The player must move through the different areas of the room, find clues and solve puzzles to find the exit.

The board is composed of three areas: kitchen, living room and bedroom. Each area has its own puzzle that must be solved to progress in the game. The player can move between the areas by typing the name of the desired area. When the player is in the area, he can see the information about the current puzzle, including the clue.

The player can try to solve the puzzle by typing the solution. If the solution is correct, the puzzle is removed and the player can continue to the next area. Otherwise, the player must try again.

The game also has a timer, with an initial time of 60 minutes. Each turn, the remaining time is decremented by one minute. If the player manages to escape before the time runs out, he wins the game. Otherwise, he loses.

This game is a basic implementation and can be expanded and improved with more features and game mechanics. For example, more puzzles, more areas, and the ability to acquire and use items to help solve the puzzles can be added. In addition, a graphical interface can be included to make the game more attractive.